add_operation

+++ prismtest2/Models/Entity/IPerson.cs
@@ -8,7 +8,7 @@ namespace prismtest2.Models.Entity
 {
     public interface IPerson
     {
-        public UInt64 Id { get; set; }
+        public int Id { get; set; }
         public string Username { get; set; }
         public string Password { get; set; }
         public string Email { get; set; }
