+++ prismtest2/Models/Entity/Users.cs
@@ -8,7 +8,7 @@ namespace prismtest2.Models.Entity
 {
     public class Users:IPerson
     {
-        public UInt64 Id { get; set; }
+        public int Id { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
         public string Username { get; set; }
