+++ prismtest2/Models/Services/ICreateUser.cs
@@ -12,7 +12,7 @@ namespace prismtest2.Models.Services
     }
     public class UserListDTO
     {
-        public UInt64 id;
+        public int id;
         public string User_name;
         public string Password;
         public string Email;
